Allow CLI and Cavlcade to run in Codeception tests
Related to https://github.com/humanmade/altis-dev-tools/pull/248

I was having all kinds of issues with cron jobs running on the main db etc... This is by no means a perfect method using a file as an indicator but I tried all kinds of other approaches like exporting different constants etc on the container but this is only thing that worked.

Also adds in some additional clean up for ES indexes and uses a shutdown hook to ensure test dbs and the `.test-running` file are cleaned up.