Flush cache before intiating test mode
I hit some odd behaviour where the database could not be reached, turned out it was due to the cache not being cleared already when calling `wp cache flush` in the loop to run suites. This makes it run for the actual local site, and then again after each test run.